voice-assistant/requirements.txt
# --- gRPC Core ---
grpcio==1.62.0
grpcio-tools==1.62.0

# --- Networking & Async ---
# Used for async calls to Ollama LLM
httpx==0.27.0

# --- AI Engines ---
# Speach-to-Text: Pinning ensures model compatibility
faster-whisper==1.0.3
# Text-to-Speech: Required for Piper ONNX execution
piper-tts==1.2.0

# --- Observability & Tracing ---
# Mandatory for cross-service tracing (Go -> Python)
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-grpc

# --- Utilities ---
# Faster math for audio processing
numpy<2.0.0